Report unknown attributes as an RFC 7807 "Problem Detail"
When we implement multiple levels of authentication we will need to
return structured errors which communicate, at least, which level of
authentication a user needs to be re-authenticated at for their
request to succeed.

If we're introducing structured errors, why not adopt an existing
approach?

RFC 7807[1] is a proposed standard for a JSON format of returning
errors from APIs.  A "problem detail" has three parts:

- the "type", which is a pointer to further documentation (I really
  like this)
- the "title", which names this class of problem
- the "detail", which describes this specific problem

And any extra fields, to include (eg) machine-readable resolution
details.

[1] https://tools.ietf.org/html/rfc7807